python_app/requirement.txt
paho-mqtt
mysql-connector-python
pymongo
neo4j
